pyproject.toml
[project]
name = "agent-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.29.0",
    "ipykernel>=6.29.5",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
